all primitive model creators with resolution are behaving weirdly and I couldn't figure it out. but as long as the shape is closed and well builded, mesh creation is working fine. if the shape is not closed, there will be some boundary edges and their right faces will be nullptr.

+++ GamEncin/src/Encin/Tools/MathYaman.cpp
@@ -6,10 +6,10 @@ namespace GamEncin
     {
 #pragma region Constants
 
-        const float E = 2.7182817F,
-            PI = 3.1415927F,
-            GRAVITY = 9.80665F,
-            G = 6.67F / 1e11F;
+        const double E = 2.7182817,
+            PI = 3.1415927,
+            GRAVITY = 9.80665,
+            G = 6.67 / 1e11;
 
 #pragma endregion
